Multiple reference genomes, WS API and documentation, External raw files links and downloads

+++ src/org/irri/iric/portal/chado/dao/VIricstocksByPhenotypeDAO.java
@@ -31,13 +31,13 @@ public interface VIricstocksByPhenotypeDAO extends
 	 * JPQL Query - findVIricstocksByPhenotypeByPrimaryKey
 	 *
 	 */
-	public VIricstocksByPhenotype findVIricstocksByPhenotypeByPrimaryKey(Integer iricStockPhenotypeId) throws DataAccessException;
+	public VIricstocksByPhenotype findVIricstocksByPhenotypeByPrimaryKey(BigDecimal iricStockPhenotypeId) throws DataAccessException;
 
 	/**
 	 * JPQL Query - findVIricstocksByPhenotypeByPrimaryKey
 	 *
 	 */
-	public VIricstocksByPhenotype findVIricstocksByPhenotypeByPrimaryKey(Integer iricStockPhenotypeId, int startResult, int maxRows) throws DataAccessException;
+	public VIricstocksByPhenotype findVIricstocksByPhenotypeByPrimaryKey(BigDecimal iricStockPhenotypeId, int startResult, int maxRows) throws DataAccessException;
 
 	/**
 	 * JPQL Query - findVIricstocksByPhenotypeByOriCountry
